feature(authentication): standardise code format & prepare for deployment (#1)
This doesn't provide any new functionality. It is simply a tidy-up and standardisation of the code.

Add linting and configuration as implemented in the "root" repo.
Simplify authentication code and tests.
Remove usage of SCSS modules as we are going to standardise on not using them in any of our repos.